feat(#678): move document aside detail pages
