+++ apps/web/src/app/[lang]/dho/[id]/@aside/agreements/[documentSlug]/page.tsx
@@ -1,11 +1,10 @@
 'use client';
-import { AgreementDetail } from '@hypha-platform/epics';
-import { SidePanel } from '../../../@aside/_components/side-panel';
+import { AgreementDetail, SidePanel } from '@hypha-platform/epics';
 import { useParams } from 'next/navigation';
-import { getDhoPathAgreements } from '../../constants';
 import { Locale } from '@hypha-platform/i18n';
 import { useDocumentBySlug } from '@web/hooks/use-document-by-slug';
 import { useDocumentSlug } from '@web/hooks/use-document-slug';
+import { getDhoPathAgreements } from '../../../@tab/agreements/constants';
 
 export default function Agreements() {
   const { id, lang } = useParams();
